:art: Improved the structure of code and seperated components. Also added an error if a city isn't found in search
