+++ src/components/ReactRain/index.js
@@ -1,17 +1,72 @@
-import React from "react";
-import PageLayout from "../../layout/pageLayout";
-import {Title,Copy} from "./styled.js"
-import WeatherApp from "../GeoLocation";
+import React, { useState, useEffect } from 'react';
+import SearchBar from '../SearchBar';
+import WeatherDisplay from '../WeatherDisplay';
+
+const apiKey = process.env.REACT_APP_WEATHER_API_KEY;
+const apiUrl = 'https://api.openweathermap.org/data/2.5/weather';
+const geoApiUrl = 'https://api.openweathermap.org/geo/1.0/direct';
 
 const ReactRain = () => {
+  const [weatherData, setWeatherData] = useState(null);
+  const [cityNotFound, setCityNotFound] = useState(false);
 
-  
+  const fetchWeatherByLatLng = async (lat, lon) => {
+    try {
+      const response = await fetch(`${apiUrl}?lat=${lat}&lon=${lon}&appid=${apiKey}&units=metric`);
+      if (response.ok) {
+        const data = await response.json();
+        setWeatherData(data);
+        setCityNotFound(false); // Reset city not found error if data is fetched.
+      } else {
+        setCityNotFound(true); // Set city not found error to true.
+        console.error('Failed to fetch weather data');
+      }
+    } catch (error) {
+      console.error('Error while fetching weather data', error);
+    }
+  };
+
+  useEffect(() => {
+    if (navigator.geolocation) {
+      navigator.geolocation.getCurrentPosition((position) => {
+        const { latitude, longitude } = position.coords;
+        fetchWeatherByLatLng(latitude, longitude);
+      });
+    }
+  }, []);
+
+  const handleCitySearch = async (city) => {
+  try {
+    const geoResponse = await fetch(`${geoApiUrl}?q=${city}&limit=5&appid=${apiKey}&type=city`);
+    if (geoResponse.ok) {
+      const geoData = await geoResponse.json();
+      if (geoData.length > 0) {
+        const firstCity = geoData[0];
+        fetchWeatherByLatLng(firstCity.lat, firstCity.lon);
+      } else {
+        setCityNotFound(true);
+        console.error('City not found');
+      }
+    } else {
+      setCityNotFound(true);
+      console.error('Failed to fetch city data');
+    }
+  } catch (error) {
+    console.error('Error while fetching city data', error);
+  }
+};
+
 
   return (
-    <PageLayout>
-<Title>Hello World.</Title>
-<Copy><WeatherApp/></Copy>
-    </PageLayout>
+    <div>
+      <h1>Weather App</h1>
+      <SearchBar onSearch={handleCitySearch} />
+      {cityNotFound ? (
+        <div>City not found</div>
+      ) : (
+        <WeatherDisplay weatherData={weatherData} />
+      )}
+    </div>
   );
 };
